New classes for CCS prediction and molecule representation
A new class `CcsPredictionModel` has been added to the file `CcsPredictionModel.cs`. This class uses ML.NET to train, evaluate and predict a collisional cross section (CCS) prediction model. It contains methods for training (`TrainAsync`), predicting (`PredictAsync`), evaluating (`EvaluateAsync`), saving (`SaveModel`) and loading (`LoadModel`) the model.

A new class `CcsPredictor` has been added in the file `CcsPredictor.cs`. This class uses `CcsPredictionModel` to train and evaluate the model and make CCS predictions. It contains methods for training and evaluating the model (`TrainAndEvaluateAsync`) and for predicting CCS values (`PredictCcsAsync`).

A new class `Molecule` was added in the file `Molecule.cs`. This class represents a molecule with SMILES and InChI representations and a collection of descriptors.

A new class `MoleculeWithCcs` has been added in the file `MoleculeWithCcs.cs`. This class represents a molecule with SMILES and InChI representations and a CCS value.