Expansion of data providers and ONNX integration
The changes improve the flexibility and functionality of the code by adding support for multiple data providers and additional fingerprint types. Data preparation has been optimized and made asynchronous to improve performance. Model classes have been adapted to use the new descriptor calculator objects and molecular data, which should increase the accuracy and efficiency of predictions.

A new project called `CCSPredict.OnnxLoader` has been added to the `CCSPredict.sln` solution. This project uses the .NET SDK and targets .NET 8.0. It contains a package reference to the `Microsoft.ML.OnnxRuntime` library. A new `Program` class has been added that uses the `Microsoft.ML.OnnxRuntime` library to make predictions based on ONNX models. Added a helper class `OnnxModel` that loads an ONNX model and provides a method `Predict` to make predictions based on input data. A placeholder method `CalculateDescriptors` has been added, which should be replaced by an actual calculation of the descriptors in a real application.

In addition, a new class `TaskExtension` has been added to the `CCSPredict.Core` namespace, which contains a static method `Await` to synchronously wait for a `Task` instance. The `CCSPredict.sln` file has been updated to include the new `CCSPredict.OnnxLoader` project and its configurations for debug and release.